+++ core/domain/src/main/java/com/ravikantsharma/core/domain/model/DecimalSeparator.kt
@@ -1,12 +1,13 @@
 package com.ravikantsharma.core.domain.model
 
+import java.math.BigDecimal
 import java.util.Locale
 
 enum class DecimalSeparator : PreferenceOption {
     DOT,
     COMMA;
 
-    override fun displayText(number: Double, currency: Currency?, keepDecimal: Boolean): String {
+    override fun displayText(number: BigDecimal, currency: Currency?, keepDecimal: Boolean): String {
         return when (this) {
             DOT -> String.format(Locale.US, "%.2f", number)
             COMMA -> String.format(Locale.US, "%.2f", number).replace('.', ',')
